Replace handlers with return values
Instead of calling handlers when certain events occur inside the VM, we
suspend and return an enum when the events occur. The enum contains the
data that would go in the handlers. Then the client can decide how to
handle the data and whether to continue the VM or not.